move ios openUrl fix to separate file
