+++ packages/auth/plugin/src/index.ts
@@ -1,6 +1,6 @@
 import { ConfigPlugin, withPlugins, createRunOncePlugin } from '@expo/config-plugins';
 
-import { withIosCaptchaUrlTypes } from './ios';
+import { withIosCaptchaUrlTypes, withIosCaptchaOpenUrlFix } from './ios';
 
 /**
  * A config plugin for configuring `@react-native-firebase/auth`
@@ -9,6 +9,7 @@ const withRnFirebaseAuth: ConfigPlugin = config => {
   return withPlugins(config, [
     // iOS
     withIosCaptchaUrlTypes,
+    withIosCaptchaOpenUrlFix,
   ]);
 };
